Testing and fixing Hetzner VPS template build
